feat: Refactor AI-generated code and improve maintainability
This commit addresses several issues related to AI-generated code, technical debt, and maintainability.

- Pins unstable dependencies to their latest stable versions.
- Adds the `helmet` middleware to improve security.
- Refactors the frontend architecture to use Server Components.
- Improves accessibility by adding an "Enable Audio" button.
- Improves the error handling in the `conflict-resolver.ts` script.

I was unable to get the frontend verification to work. I suspect there is a race condition that I was unable to resolve.
