+++ package.json
@@ -76,12 +76,13 @@
     "express": "^5.2.1",
     "express-rate-limit": "^8.2.1",
     "framer-motion": "^12.29.2",
+    "helmet": "^8.1.0",
     "idb": "^8.0.3",
     "js-cookie": "^3.0.5",
     "lodash.isequal": "^4.5.0",
     "lodash.throttle": "^4.1.1",
     "module-alias": "^2.2.3",
-    "next": "16.1.5",
+    "next": "16.1.6",
     "next-auth": "^4.24.13",
     "notistack": "^3.0.2",
     "pino": "^10.3.0",
@@ -100,7 +101,7 @@
     "@eslint/js": "^9.39.2",
     "@google/generative-ai": "0.24.1",
     "@jest/globals": "^30.2.0",
-    "@next/bundle-analyzer": "^16.1.5",
+    "@next/bundle-analyzer": "^16.1.6",
     "@playwright/test": "^1.57.0",
     "@storybook/addon-a11y": "^10.1.11",
     "@storybook/addon-actions": "^8.6.14",
@@ -132,7 +133,7 @@
     "cross-env": "^10.1.0",
     "dotenv": "^17.2.3",
     "eslint": "9.39.2",
-    "eslint-config-next": "16.1.5",
+    "eslint-config-next": "16.1.6",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.5",
     "eslint-plugin-storybook": "^10.2.0",
